requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2
python-dotenv
drf-nested-routers